pyproject.toml
[tool.poetry]
name = "python-flask"
version = "0.1.0"
description = "Start a simple web app with the Flask framework for Python"
authors = ["Google LLC"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^3.0.1"
python-dotenv = "^1.1.0"
langchain_core = "^0.3.51"
langchain-google-genai = "^2.1.2"
neo4j = "^5.28.1"
langchain-neo4j = "^0.4.0"
numpy = "^2.2.4"

[tool.poetry.scripts]
app = 'src.main:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
